+++ app/src/main/java/com/example/passit/rvadapters/SubjectsViewRVAdapter.java
@@ -1,6 +1,8 @@
 package com.example.passit.rvadapters;
 
 import android.annotation.SuppressLint;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Paint;
 import android.os.Bundle;
@@ -8,6 +10,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.Filter;
+import android.widget.Filterable;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -17,19 +21,25 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.passit.AppDatabase;
+import com.example.passit.NotificationSender;
 import com.example.passit.R;
 import com.example.passit.SubjectDetails;
+import com.example.passit.db.entities.Notification;
+import com.example.passit.db.entities.Responsibility;
 import com.example.passit.db.entities.Subject;
 
+import java.util.ArrayList;
 import java.util.List;
 
-public class SubjectsViewRVAdapter extends RecyclerView.Adapter<SubjectsViewRVAdapter.ViewHolder> {
+public class SubjectsViewRVAdapter extends RecyclerView.Adapter<SubjectsViewRVAdapter.ViewHolder> implements Filterable {
 
     private final List<Subject> subjectNameList1;
+    private final List<Subject> subjectNameListFull;
     private AppDatabase db;
 
     public SubjectsViewRVAdapter(List<Subject> subjectNameList1) {
         this.subjectNameList1 = subjectNameList1;
+        subjectNameListFull = new ArrayList<>(subjectNameList1);
     }
 
     @NonNull
@@ -101,10 +111,36 @@ public class SubjectsViewRVAdapter extends RecyclerView.Adapter<SubjectsViewRVAd
         holder.removeBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                db.profileDao().deleteSubject(subjectId);
-                subjectNameList1.remove(holder.getAdapterPosition());
-                notifyItemRemoved(holder.getAdapterPosition());
-                notifyItemRangeChanged(holder.getAdapterPosition(), subjectNameList1.size());
+
+                AlertDialog.Builder builder = new AlertDialog.Builder(holder.subjectName.getContext(), AlertDialog.THEME_DEVICE_DEFAULT_DARK);
+                builder.setTitle("Usuń przedmiot");
+                builder.setMessage("Czy na pewno chcesz usunąć przedmiot " + subjectNameList1.get(holder.getAdapterPosition()).getSubject_name() + "? Usunięcie przedmiotu usunie wszystkie przypisane do niego zadania!");
+                builder.setCancelable(false);
+                builder.setPositiveButton("Tak", (DialogInterface.OnClickListener) (dialog, which) -> {
+                    NotificationSender notificationSender = new NotificationSender(holder.subjectName.getContext());
+                    List<Responsibility> responsibilities = db.profileDao().getResponsibilityWithSubjectId(subjectNameList1.get(holder.getAdapterPosition()).subject_id);
+
+                    for (Responsibility resp : responsibilities) {
+                        notificationSender.cancelNotification(db.profileDao().getNotificationId(resp.getResp_id(), "Reminder"));
+                        notificationSender.cancelNotification(db.profileDao().getNotificationId(resp.getResp_id(), "Delay"));
+                        db.profileDao().deleteNotificationById(db.profileDao().getNotificationId(resp.getResp_id(), "Reminder"));
+                        db.profileDao().deleteNotificationById(db.profileDao().getNotificationId(resp.getResp_id(), "Delay"));
+                        db.profileDao().deleteResponsibility(resp.getResp_id());
+                    }
+
+                    db.profileDao().deleteSubject(subjectId);
+                    subjectNameList1.remove(holder.getAdapterPosition());
+                    notifyItemRemoved(holder.getAdapterPosition());
+                    notifyItemRangeChanged(holder.getAdapterPosition(), subjectNameList1.size());
+                    dialog.dismiss();
+                });
+
+                builder.setNegativeButton("Nie", (DialogInterface.OnClickListener) (dialog, which) -> {
+                    dialog.cancel();
+                });
+
+                AlertDialog alertDialog = builder.create();
+                alertDialog.show();
             }
         });
 
@@ -124,6 +160,41 @@ public class SubjectsViewRVAdapter extends RecyclerView.Adapter<SubjectsViewRVAd
     public int getItemCount() {
         return subjectNameList1.size();
     }
+
+    @Override
+    public Filter getFilter() {
+        return respFilter;
+    }
+
+    private Filter respFilter = new Filter() {
+        @Override
+        protected FilterResults performFiltering(CharSequence charSequence) {
+            List<Subject> filteredList = new ArrayList<>();
+
+            if (charSequence == null || charSequence.length() == 0) {
+                filteredList.addAll(subjectNameListFull);
+            } else {
+                String filterPattern = charSequence.toString().toLowerCase().trim();
+
+                for (Subject subject : subjectNameListFull) {
+                    if (subject.getSubject_name().toLowerCase().contains(filterPattern)) {
+                        filteredList.add(subject);
+                    }
+                }
+            }
+            FilterResults results = new FilterResults();
+            results.values = filteredList;
+            return results;
+        }
+
+        @SuppressLint("NotifyDataSetChanged")
+        @Override
+        protected void publishResults(CharSequence charSequence, FilterResults results) {
+            subjectNameList1.clear();
+            subjectNameList1.addAll((List) results.values);
+            notifyDataSetChanged();
+        }
+    };
 
     public class ViewHolder extends RecyclerView.ViewHolder {
         private final TextView subjectName, progressTV;
